requirements.txt
playwright==1.47.0
Pillow==10.4.0
python-dateutil==2.9.0.post0
tqdm==4.66.4
rich==13.7.1
PySide6==6.7.2
psutil==5.9.8
requests==2.32.3
